Splitting pagination grouping functionality into its own file
